refactor: get stations wrapper, calendar date
First off, I must have been five per mille drunk when I wrote the calendar date utility, simplify it, a lot.
Now, instead of using the filesystem as database just memoize the fetch using parameters as keys. This makes the functionality a lot clearer
