--- site/src/lib/digitraffic/server/index.ts
@@ -1,1 +1,0 @@
-export * from './get_stations'
